Finish page layout
I had to make some compromises on how I wanted to finish this
page out, but that's just how it goes sometimes. I'd rather
finish something and not be entirely happy with it rather
than never finish anything at all. I can always come back to it
later, but now is the time to keep moving forward.